requirements.txt
supabase==2.13.0  # Latest version you installed
google-generativeai==0.3.2
python-telegram-bot==20.7
flask==3.0.0
langdetect==1.0.9
pytest==7.4.0
httpx==0.25.2  # Match python-telegram-bot requirement